Change ulong casts to literals

+++ Rudim.Test/BitboardTest.cs
@@ -10,7 +10,7 @@ namespace Rudim.Test
             var Board = new Bitboard(0);
 
             Board.SetBit(5);
-            Assert.Equal((ulong)32, Board.Board);
+            Assert.Equal(32ul, Board.Board);
 
             Board.SetBit(63);
             Assert.Equal(9223372036854775840, Board.Board);
@@ -22,10 +22,10 @@ namespace Rudim.Test
             var Board = new Bitboard(9223372036854775840);
 
             Board.ClearBit(63);
-            Assert.Equal((ulong)32, Board.Board);
+            Assert.Equal(32ul, Board.Board);
 
             Board.ClearBit(5);
-            Assert.Equal((ulong)0, Board.Board);
+            Assert.Equal(0ul, Board.Board);
         }
 
         [Fact]
@@ -34,10 +34,10 @@ namespace Rudim.Test
             var Board = new Bitboard(0);
 
             Board.SetBit(63);
-            Assert.Equal(9223372036854775808, Board.Board);
+            Assert.Equal(9223372036854775808ul, Board.Board);
 
             Board.SetBit(63);
-            Assert.Equal(9223372036854775808, Board.Board);
+            Assert.Equal(9223372036854775808ul, Board.Board);
         }
 
         [Fact]
@@ -46,10 +46,10 @@ namespace Rudim.Test
             var Board = new Bitboard(9223372036854775808);
 
             Board.ClearBit(63);
-            Assert.Equal((ulong)0, Board.Board);
+            Assert.Equal(0ul, Board.Board);
 
             Board.ClearBit(63);
-            Assert.Equal((ulong)0, Board.Board);
+            Assert.Equal(0ul, Board.Board);
         }
 
         [Fact]
